+++ CHANGELOG.md
@@ -64,3 +64,4 @@ Initial version
 * Added `PHPStan` to `composer.json` to a dev environment.
 * Updated Customer entity to support cross usage.
 * Updated Event entity to support cross usage.
+* Updated Transaction entity to support cross usage.